@@ -1,36 +1,37 @@
 {
   "name": "webauthn-codelab",
-  "version": "0.0.3",
+  "version": "1.0.0",
   "description": "Your First WebAuthn App",
-  "main": "server.js",
+  "main": "server.mjs",
+  "type": "module",
   "scripts": {
-    "build": "webpack -p",
-    "start": "node server.js"
+    "build": "webpack",
+    "start": "node server.mjs"
   },
   "dependencies": {
-    "express": "^4.17.1",
-    "hbs": "^4.1.1",
-    "lowdb": "^1.0.0",
-    "@material/mwc-button": "^0.25.2",
-    "@material/mwc-fab": "^0.25.2",
-    "@material/mwc-top-app-bar-fixed": "^0.25.2",
-    "@material/mwc-textfield": "^0.25.2",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "@babel/core": "^7.12.3",
-    "babel-loader": "^8.1.0",
+    "express": "^4.18.2",
+    "hbs": "^4.2.0",
+    "lowdb": "^5.1.0",
+    "@material/mwc-button": "^0.27.0",
+    "@material/mwc-fab": "^0.27.0",
+    "@material/mwc-top-app-bar-fixed": "^0.27.0",
+    "@material/mwc-textfield": "^0.27.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "@babel/core": "^7.20.12",
+    "babel-loader": "^9.1.2",
     "babel-preset-env": "^1.7.0",
-    "@simplewebauthn/server": "^4.1.0",
+    "@simplewebauthn/server": "^7.0.1",
     "base64url": "^3.0.1",
-    "css-loader": "^4.3.0",
-    "sass-loader": "^10.0.4",
-    "sass": "^1.27.2",
+    "css-loader": "^6.7.3",
+    "sass-loader": "^10.4.1",
+    "sass": "^1.58.3",
     "extract-loader": "^5.1.0",
-    "file-loader": "^6.1.1",
-    "express-session": "^1.17.2"
+    "file-loader": "^6.2.0",
+    "express-session": "^1.17.3"
   },
   "engines": {
-    "node": "10.x"
+    "node": "16.x"
   },
   "repository": {
     "url": "https://github.com/googlecodelabs/fido2-codelab"
