@@ -1,4 +1,22 @@
-module.exports = {
+/*
+ * @license
+ * Copyright 2023 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+import sass from 'sass';
+
+export default {
   entry: {
     'components-bundle': './public/components.js',
     'styles-bundle': './public/style.scss',
@@ -23,7 +41,7 @@ module.exports = {
           {
             loader: 'sass-loader',
             options: {
-              implementation: require('sass'),
+              implementation: sass,
               webpackImporter: false,
               sassOptions: {
                 includePaths: ['./node_modules'],
@@ -34,8 +52,6 @@ module.exports = {
       },
       {
         test: /components\.js$/,
-        loader: 'babel-loader',
-        query: { presets: ['env'] },
       },
     ],
   },
